Rename utils.py to swaps.py
